README.md
the link to the index.html does not become a webpage. Ok. Just point to the (?) button in the extension for further info.

(Note to future self: the index.html is handedited. It can't be converted back to markdown. This readme can't be html. To prevent lengthy manual editing, I converted the introduction to md only.)